Added PIA-Connect v1.1
This workflow utilises the PIA CLI piactl to allow you to connect
to a PIA VPN region using Alfred. It requires that the GUI is
running at this time, due to a limitation of PIA.app, but options
to address this may be explored at a later date.

Anyway, the workflow allows the user to connect to a specific
region, the last region that was connected to or the 'best'
region, as defined by PIA. It can be used to disconnect from the
VPN and authenticate/logout of the VPN.

So far, v1.1 works as far as I've tested. It's only been tested on
my local device so here's hoping it works wherever I clone this to!